+++ core-api/src/main/kotlin/org/studieojavry/coreapi/errorcase/case/presentation/web/ErrorCaseController.kt
@@ -275,6 +275,7 @@ class ErrorCaseController(
                     paste = request.paste,
                     description = request.description,
                     severityCode = request.severity,
+                    tags = request.tags,
                     snippetMarkerIds = request.snippetMarkerIds,
                     attachmentMarkerIds = request.attachmentMarkerIds,
                     status = request.status,